MySQL 依存部分を分離
