+++ src/server/Model/OperatorModelFactorySetting.ts
@@ -1,9 +1,9 @@
 import factory from './ModelFactory';
-import { ServicesDB } from './DB/ServicesDB';
-import { ProgramsDB } from './DB/ProgramsDB';
-import { RulesDB } from './DB/RulesDB';
-import { RecordedDB } from './DB/RecordedDB';
-import { EncodedDB } from './DB/EncodedDB';
+import MySQLServicesDB from './DB/MySQL/ServicesDB';
+import MySQLProgramsDB from './DB/MySQL/ProgramsDB';
+import MySQLRulesDB from './DB/MySQL/RulesDB';
+import MySQLRecordedDB from './DB/MySQL/RecordedDB';
+import MySQLEncodedDB from './DB/MySQL/EncodedDB';
 import { IPCServer } from './IPC/IPCServer';
 import { ExternalProcessModel } from './ExternalProcessModel';
 
@@ -15,14 +15,15 @@ namespace ModelFactorySetting {
     * Model をセットする
     */
     export const init = (): void => {
-        factory.reg('ServicesDB', () => { return new ServicesDB() });
-        factory.reg('ProgramsDB', () => { return new ProgramsDB() });
-        factory.reg('RulesDB', () => { return new RulesDB() });
-        factory.reg('RecordedDB', () => { return new RecordedDB() });
-        factory.reg('EncodedDB', () => { return new EncodedDB() });
+        factory.reg('ServicesDB', () => { return new MySQLServicesDB() });
+        factory.reg('ProgramsDB', () => { return new MySQLProgramsDB() });
+        factory.reg('RulesDB', () => { return new MySQLRulesDB() });
+        factory.reg('RecordedDB', () => { return new MySQLRecordedDB() });
+        factory.reg('EncodedDB', () => { return new MySQLEncodedDB() });
         factory.reg('IPCServer', () => { return IPCServer.getInstance(); });
         factory.reg('ExternalProcessModel', () => { return new ExternalProcessModel(); });
     }
 }
 
 export default ModelFactorySetting;
+
